Added earth_imagery method
earth_imagery returns a tuple containing a NumPy array with the pixel values of the image matching given data and the URL to that image. For easier visualization, the display parameter when True opens a cv2 window displaying that image and the save_as when it's not None saves the image to a .png file with the specified name.